Fixing Canvas-bounds encoding

+++ src/me/anno/ui/canvas/Bounds.kt
@@ -1,3 +1,3 @@
 package me.anno.ui.canvas
 
-data class Bounds(val x0: Int, val y0: Int, val x1: Int, val y1: Int)
+data class Bounds(val x0: Short, val y0: Short, val x1: Short, val y1: Short)
